+++ musicbot/cogs/cogmanager.py
@@ -4,8 +4,8 @@ from discord.ext import commands
 import traceback
 
 # Bot Utilities
-from cogs.helpformatter import commandhelper
-from cogs.utils.userinteraction import Scroller
+from .helpformatter import commandhelper
+from ..utils.userinteraction import Scroller
 
 
 class CogManager(commands.Cog):
