+++ musicbot/cogs/helpformatter/__init__.py
@@ -1,0 +1,8 @@
+from .help import Help
+from .helpformatter import commandhelper
+
+__all__ = ['Help', 'commandHelper']
+
+def setup(bot):
+    bot.remove_command("help")
+    bot.add_cog(Help(bot))
