+++ musicbot/cogs/settings.py
@@ -3,9 +3,9 @@ import discord
 from discord.ext import commands
 
 # Bot Utilities
-from cogs.helpformatter import commandhelper
-from cogs.utils import checks
-from cogs.utils.userinteraction import Scroller
+from .helpformatter import commandhelper
+from ..utils import checks
+from ..utils.userinteraction import Scroller
 
 
 class Settings(commands.Cog):
