+++ musicbot/utils/bot_version.py
