+++ musicbot/utils/checks.py
