+++ musicbot/utils/thumbnailer.py
