+++ musicbot/utils/timeformatter.py
