+++ musicbot/utils/userinteraction/paginators/textpaginator.py
